new changes to fantasy league

+++ frontend/src/components/leagues/PredictionLeague/GlobalCards.jsx
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import Badges from '../../../images/badges/exportBadges.js'; // Adjust the path as per your project structure
 import axios from 'axios';
+import { toast } from 'react-toastify';
 import '../../../styles/cards.css'
 
-function TeamCard({ gameId, roundNum, team1Name, matchDate, matchTime, team2Name }) {
+function TeamCard({ gameId, roundNum, team1Name, matchDate, matchTime, team2Name, blockChanges }) {
   // State to manage whether the prediction button is clicked
   const [isPredicted, setIsPredicted] = useState(false);
   // State to manage the input values for the match score
@@ -51,12 +52,20 @@ function TeamCard({ gameId, roundNum, team1Name, matchDate, matchTime, team2Name
   };
 
   const handlePredictionClick = () => {
+    if(blockChanges){
+      toast.error(`Round ${roundNum} Predictions Window Closed `);
+      return
+    }
     setIsPredicted(true);
   }; 
 
 
   // Function to handle cancel prediction
   const handleCancelPrediction = async () => {
+    if(blockChanges){
+      toast.error(`Round ${roundNum} Predictions Window Closed `);
+      return
+    }
     try {
       if (isPredicted) {
         const response = await axios.delete(`http://localhost:3000/api/deleteGlobalPrediction/${gameId}`, {
@@ -86,7 +95,12 @@ function TeamCard({ gameId, roundNum, team1Name, matchDate, matchTime, team2Name
 
   // Function to handle save prediction
   const handleSavePrediction = async () => {
-    // Check if both scores are provided
+
+    if(blockChanges){
+      toast.error(`Round ${roundNum} Predictions Window Closed `);
+      return
+    }
+
     if (!team1Score || !team2Score) {
       console.warn('Please enter scores for both teams.');
       return;
@@ -202,7 +216,7 @@ function TeamCard({ gameId, roundNum, team1Name, matchDate, matchTime, team2Name
   );
 }
 
-function Card({ gamePairs }) {
+function Card({ gamePairs, blockChanges }) {
   return (
     <div className="container">
       <div className="teams-card-container">
@@ -217,6 +231,8 @@ function Card({ gamePairs }) {
               team2Name={pair.team_2}
               roundNum={pair.round_num} 
               gameId={pair.game_id}
+              blockChanges={blockChanges}
+
             />
           );
         })}
